chore: create `record_batch_errors` module within `arrow` module
