+++ crates/proof-of-sql/src/base/arrow/record_batch_errors.rs
@@ -1,0 +1,38 @@
+use super::arrow_array_to_column_conversion::ArrowArrayToColumnConversionError;
+use crate::base::commitment::ColumnCommitmentsMismatch;
+use proof_of_sql_parser::ParseError;
+use snafu::Snafu;
+
+/// Errors that can occur when trying to create or extend a [`TableCommitment`] from a record batch.
+#[derive(Debug, Snafu)]
+pub enum RecordBatchToColumnsError {
+    /// Error converting from arrow array
+    #[snafu(transparent)]
+    ArrowArrayToColumnConversionError {
+        /// The underlying source error
+        source: ArrowArrayToColumnConversionError,
+    },
+    #[snafu(transparent)]
+    /// This error occurs when convering from a record batch name to an identifier fails. (Which may be impossible.)
+    FieldParseFail {
+        /// The underlying source error
+        source: ParseError,
+    },
+}
+
+/// Errors that can occur when attempting to append a record batch to a [`TableCommitment`].
+#[derive(Debug, Snafu)]
+pub enum AppendRecordBatchTableCommitmentError {
+    /// During commitment operation, metadata indicates that operand tables cannot be the same.
+    #[snafu(transparent)]
+    ColumnCommitmentsMismatch {
+        /// The underlying source error
+        source: ColumnCommitmentsMismatch,
+    },
+    /// Error converting from arrow array
+    #[snafu(transparent)]
+    ArrowBatchToColumnError {
+        /// The underlying source error
+        source: RecordBatchToColumnsError,
+    },
+}
